Refactor code so the calls to the api are async actions using thunk
